Revert r82120: untested code causes fatal error on Special:Contributions when using CentralAuth extension.
The hook is receiving a Title object, then passing it to a function that expects a User object. This naturally fails with a fatal error when User-specific methods are called.
Please don't make commits to CentralAuth without testing them; it's an important component to production sites and it's easy to break it. If it's worth making the change, it's worth setting up a test environment. :)