(bug 20272) Make AbortAutoAccount useful for other extensions.
Before this change AbortAutoAccount was called on empty User object,
so the extensions were unable to even check the user name of created
account.

This change moves User object initialisation before hook is called,
so User object is at least populated with user name. There is also
additional check if the user name was changed by one of the hooks.

Custom CentralAuthAutoCreate is now marked as deprecated as extensions
should use generic AbortAutoAccount.

If autocreation is stopped by a hook, the wiki is added to blacklist,
so expensive checks won't be run on every page refresh. According to a
comment in the code, the user is still able to use Special:UserLogin.

Change-Id: I92f57fc2c3189c42157478de14a8d48045a253b5