(bug 14759) Add support for global user pages.
This change adds support for global user pages. It heavily uses caching.
The user page is fetched using http request from the global wiki if
local user page does not exist. The content is stored in the cache.

There is performance issue regarding marking red links to user pages
as blue. Any constructive feedback is welcome.

I am not sure if this feature should be included in this extension
or a separate extension should be created.

Change-Id: Idc09268fc4a1bcd25b0faee3d2a97ed25f2f90d0