Update patch set 3
Patch Set 3:

Thanks Krinkle for your review. ResourceLoader does not have a clear documentation. So if I can avoid near empty class, please tell me how to do it :).

I have added a note to the class documentation what is the role of that module. I hope it is now clear enough.

Patch-set: 3