Update patch set 4
Patch Set 4: Do not submit

The empty class to use on local wikis for registering a module with a custom load-source is good, in general. Maybe Roan knows a better way, I think there should be.

However I'm marking this -2 for now due to lack of cache control. This class to implement getModifiedTime() some how. Since CentralAuth is granted to have database access, and the global-user-module wikis is always a wiki that is part of the db-cluster, this should be relatively easy to do.

I see that the configuration variable has already been turned into an array that takes wiki-id (in addition to load-source), which you'll need to get that information.

Patch-set: 4