Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This looks OK to me and in this case I do not think you need to worry about code duplication since it's a small amount of code that essentially represents a totally different sql query. The alternative, I guess, would be to refactor the code in UsersPager, but that seems silly to me for this particular case. Since I'm unfamiliar with CentralAuth, I will refrain from +2'ing this, but this looks fine to me.

Patch-set: 1