Update patch set 1
Patch Set 1:

Other Wikimedia wikis in our case.

This is the brother of the other centralauth-account-exists-*messages  ("is already taken in the unified login system.", "it does exist in the unified login system.").

In this case it doesn't exist in SUL, but it's as if it were.

You're welcome to propose a different wording of course :)

Patch-set: 1