Update patch set 5
Patch Set 5:

And I don't like using extractRequestParams here. I tried but it just caused issues (related to whether parameters are set or not IIRC). If you really want it working with that, you'll have to do it (alternatively, revoke your -1).

As for the long lines, what are you talking about? The longest are just under 120 characters and don't even reach the end of my screen.

Patch-set: 5