Disable editing for non-global accounts
Add a global to allow a wiki to disable editing for accounts that have
not been linked to a global account. Users are still able to edit in
the NS_USER_TALK and the NS_PROJECT_TALK namespace, to coordinate with
project sysops.

This feature was requested by the Wikivoyage migration team.

Change-Id: Ia6275a6fd36db78a82f015890212e62536ed60d4