Update patch set 8
Patch Set 8:

Though I fully agree with having this feature as it's another step towards more globality I'm aware of some users who for some reasons don't yet have global accounts (mostly probably due to the fact the extension doesn't allow them to merge).

To not prevent these people from spreading towards new projects we have to find a way to bypass this (a user right, like tbooverride probably), at least as long as we can't provide them with an easy process for global renaming.

And the configuration var is to make it possible, for 3rd party users (there are some) or whenever to make it possible to restore the old functionality of this extension.

Patch-set: 8