Update patch set 8
Patch Set 8:

The cases are basically where the extension should prevent the creation of a global account  (so it's not a bug), eg. because two users use the same user name on different projects. While those problems should definitely be resolved they aren't yet fully (especially cause renaming globally still is a huge pain).

So pushing this as is might lead to unexpected issues which then could only be resolved by patching the extension. That's why we should have this configurable and at best a user right which allows overriding this (so that eg. stewards could still create accounts, just in case)

Patch-set: 8