Reorganize and add API to lock/hide global users
Add API for locking and hiding global user accounts. This is the first
step in providing a Mass-Locking feature, as part of the Admin Tools
project.

As part of this, moved the logic for locking and hiding global accounts
into the CentralAuthUser object, so that the special page and API share
as much code as possible.

Change-Id: I9d813f481c7aea5c325e3122ae926ac6ca7cc56a