Grammar and consistency in English messages
Made "accounts" plural, because they are separate, and not
one account.

In the merged accounts there's always the current account
and at least one account from another wiki, so it's always plural.

The unmerged accounts can be singular or plural.

Based on changes in wikidata.org by [[User:Riley Huntley]].

Change-Id: If202dea4dfb21bd97c930b8b771e3be107525977