Update patch set 1
Patch Set 1: There's a problem with this change, please improve

I'm not sure this is the way to go here... IMO it would be more clean to make AntiSpoof for CentralAuth optional. Furthermore you have to add a comment to the AntiSpoof file because this one needs to be updated whenever AntiSpoof is changed.

Patch-set: 1