Global user renaming
Global renaming, using a configurable object cache for locking and
tracking the local jobs. By default, uses a new objectcache table in
the centralauth database. The WMF will probably deploy this using a
Redis instance that doesn't evict keys that are not expired.

Logging is done on each wiki where a rename is performed, in addition to
a log entry on the wiki where the rename is initiated.

Bug: 14862

Change-Id: I6f8584f28e6f4ea6e5664c925cc464dd4674899f