Update patch set 3
Patch Set 3: There's a problem with this change, please improve

(1 inline comment)

Maybe we should add hooks to AbuseFilter::generateUserVars and alike functions. That would be a much cleaner way to introduce user variables as that would also work in Special:AbuseFilter/test etc. and the right context would be available.

Patch-set: 3