Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

After talking with Hoo, this change causes the same group to identified if you pass in the group parameter, or use the $par format.

Since no WMF global groups have spaces in the names, it seems like this may be the desired behavior.

Patch-set: 1