(bug 41172) Add Wikiset as AbuseFilter variable
Use AbuseFilter hooks in CentralAuth to add CentralAuth's WikiSets as
an AbuseFilter variable. WikiSets that contain the current wiki are
included in the variable's value.

Global AbuseFilter rules are meant to apply to all wikis without
exceptions. However, in the rare case where some wikis need to be
excluded from a particular rule, WikiSets can be used.

Change-Id: I5db52aad31b4abff2eb323aafa792f4af0fe6cda