Update patch set 1
Patch Set 1:

I'm about to do that, but I still think this module should be fixed too. No other module uses null for this, it's not even clear what null means and how should it be represented in the output of paramifo.

If you think null is a valid value here, how do you propose it should be represented in paramifo in XML?

* <t /> already means empty string
* <t>null</t> would mean actually providing "locked=null"
* something like <t isnull="" />? maybe, but it doesn't feel right to me

Patch-set: 1