Update patch set 1
Patch Set 1:

It is impossible for the client to supply null as input. But since PHP's array_diff and array_intersect consider null and the empty string to be the same, null happens to work to allow the client to supply the empty string. Boolean false would also work, for that matter. But IMO it would be preferable to specify the actual allowed value instead of something that just happens to work.

Looking at the code that uses the "locked" parameter, it seems that it treats "&locked=unlocked" and "&locked=" as equivalent (note that specifying "&locked=" is *not* equivalent to omitting 'locked' entirely).

Patch-set: 1