Update patch set 4
Patch Set 4:

Updated to bring permissions into line with the change I745d2520 (Need oversight permissions to hide or suppress a user).

Also in this patchset fixed a minor bug (if a user was renamed or hidden between the time the admin searched for the username and when they tried to lock, an exception was thrown).

Patch-set: 4