(bug 18057) Add conflict detection for status
Adds a hash of the current locked and hidden values for a user in the
form submission. If another admin has updated the user's status, then
the conflict will be detected and the operation aborted.

The hash can be optionally passed with an api request, to optionally
perform the conflict detection. The MD5 should be easy to calculate in
any code using the api.

Change-Id: Ia985e152ccabb77454f0eb87bbb78f0e220d5674