Update patch set 6
Patch Set 6:

Before merging this, is there a good reason why the logs link to Special:GlobalUsers instead of back to Special:GlobalGroupPermissions? I see Wikiset logs do the same thing but wondering if we should change that for rename logs.

Also, should we consider forbidding a rename to a name with spaces in it? The <space> vs _ issue causes some weird issues, so I'm wondering if we should prevent any more names from using spaces here?

Patch-set: 6