Update patch set 5
Patch Set 5:

I like that this prevents new groups from being named with a space in the name, and when all groups comply, everything seems to work well.

When there are existing groups with spaces, then this actually breaks the rename. After renaming a group that had spaces in the name, there are two versions of the group in the database (one with spaces, one with _'s), with different rights. Just for sanity, maybe we should have a maintenance script to rename any groups that currently have spaces in the name?

Patch-set: 5