Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

A couple of syntax issues.

For the logic though, I think you only want to migrate accounts from localnames that are not in localuser, instead of everything in globalnames. Also, although you can set the batchSize in the class, you'll also need to do the limiting and looping yourself to use it.

It may also be nice to have the batch size and the prefix passed in, so the person running it can tune the size, and can save time continuing if the script is killed before it finishes.

Patch-set: 5