Update patch set 6
Patch Set 6: Verified+1 Code-Review+1

Talked with pgehres about the use case, and decided this one should be updated to be a modern maintenance script, but this will not be what we use to do the WMF migration (we'll use I6f91c697706a5e4feba2125773c6c3c259a6e29e for that).

Working in test for me though, so I'll merge unless anyone else has objections.

Patch-set: 6