Use user case for global group names, fix logging
To get proper log entries I had to add an own log action
handler for global group renames and change the target
of the log entries.

Note: This breakes b/c with older log entries, but
this shouldn't be to much of an issue as this feature
is rather new.

Bug: 46631
Bug: 27551
Change-Id: I0003998612469b62583ab53c0917ad8116465a29