Update patch set 1
Patch Set 1: Code-Review-1

It seems like half a change to me, so I'm not really too confident in reviewing it. The client side still needs to be written, right? And what API action will the client request on the remote wiki, and what will it return? Maybe this could just stay in a topic branch for now, and be merged when we have the whole package?

I'm not too happy about $wgRequest->getVal('centralauthtoken') being used interchangeably with $module->getMain()->getVal('centralauthtoken'). This was done simply because the token callbacks don't get the module object as a parameter, right? That seems to be an architectural error made as recently as MW 1.20. I think you should fix it rather than live with it, say by making the module object be the first parameter in calls to types registered with ApiTokensGetTokenTypes. The unused $pageid and $title parameters to the ApiQueryInfo methods could be removed, with a closure wrapper for callbacks registered with the APIQueryInfoTokens hook to maintain backwards compatibility.

Patch-set: 1
Reviewer: Tim Starling <17@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1