Update patch set 1
Patch Set 1:

Re "half a change", the client side of this is gadgets or user scripts on the individual wikis that want to make CORS requests without having to make the user actually visit the other site first. A "client" might look something like this:

    $.ajax( {
        url: '/w/api.php',
        dataType: 'json',
        type: 'POST',
        data: { format: 'json', action: 'tokens', type: 'centralauth' }
    } )
    .done( function ( r ) {
        origin = location.protocol + '//' + location.host;
        $.ajax( {
            url: 'https://commons.wikimedia.org/w/api.php?origin=' + encodeURIComponent( origin ) + '&centralauthtoken=' + encodeURIComponent( r.tokens.centralauthtoken ),
            dataType: 'json',
            type: 'POST',
            data: { format: 'json', action: 'query', meta: 'userinfo' }
        } )
        .done( function ( r ) {
            alert( r.query.userinfo.name );
        } );
    } );

As mentioned at the meeting yesterday, this is "I know I'm centrally logged in, and I need to tell another SUL wiki that without using cookies"; it's different from "I'm not logged in locally, I want to ask the central domain if there is a central login active so I can set my first-party cookies" which is gerrit change 58924, and it's different from "I just logged in locally, let's tell the central domain!" which appears to be Aaron's gerrit change 59211.


Re $wgRequest->getVal and $module->getMain()->getVal, there's no way to get rid of $wgRequest->getVal in favor of $module->getMain()->getVal completely. It's not just API hooks that need to check it: we also check it in SetupAfterCache (to prevent using cookies), AuthPluginSetup (to load session data from memcached instead), UserLoadFromSession (ditto), UserSetCookies (to prevent setting cookies), and anything that calls CentralAuthUser::setCookie (ditto).

And, in fact, the only call to $module->getMain()->getVal is to prevent the API from issuing a warning about the centralauthtoken parameter being unused; see bug 47046. So yes, the token handling in the API needs work (bug 45199 has some more on that), but holding this up to deal with it seems to me to be what Robla calls "yak shaving".

Patch-set: 1
Label: Verified=0