Update patch set 3
Patch Set 3:

Tested and works well. I agree with hoo's comment on PS2, it seems like there should be a better way to do that. If not, then it's probably fine as is, but it would be nice to verify.

Patch-set: 3
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0