Adding the ability for CentralAuth to login a user with an altername username.
As part of the SUL finalization, we are going to need to rename a not
insignificant number of users to $username~$wiki. Since not all users
have an email address and to reduce confusion we are adding the ability
for CentralAuth to attempt to authenticate them to alternate usernames
on a failed login attempt.

Change-Id: I0b191f76714b6da1ed63b9af709b5e21ce5ee24a