Update patch set 6
Patch Set 6:

When the user visits the wiki.

Even if we don't auto-login to all wikis from the start and instead log them in when they visit another wiki. When the user does visit that wiki we need some way of knowing they're logged in on the first wiki they logged in. There won't be any cookies on the other wiki to determine that.

Patch-set: 6