Update patch set 6
Patch Set 6:

@Tim Starling:
> Why is auto-login still needed? Why not just wait until the user actually visits another wiki?

For better user experience, to avoid the user having to first load the page as a non-logged-in user and then having to reload the page to actually see it as logged-in.

This was discussed in at least one of the meetings, although it's possible that it was only the UX meeting (which wasn't friendly to your timezone) rather than one of the normal Monday/Thursday meetings we've been having.

@Daniel Friesen:
> Though when on the other wiki we'll need some way to know whether the user is logged in on another wiki. 

If I understand you correctly, see https://www.mediawiki.org/wiki/Auth_systems/SUL2 for all sorts of details, and Ib6e9cce4

Patch-set: 6
Label: Verified=0