Update patch set 7
Patch Set 7: Code-Review-1

I tried testing this with $wgCentralAuthLoginWiki left as the default of false, since it seemed like a convenient test setup. The result is rather unpleasant: the client goes into an infinite loop of iframe POST requests, like a self-inflicted DDoS. 

$data['nextState'] is C1 for the L0 form output, causing $target to be false, causing WikiMap::getForeignURL() to issue a warning and return false. The the action attribute is blank, making the form auto-submit itself back to the same URL.

If $wgCentralAuthLoginWiki is allowed to be false, the code should check for that and do something different. If it's not allowed to be false, an exception should be thrown. $target should be validated before sending to WikiMap::getForeignURL(), to avoid the user generating PHP warnings. And the return value of WikiMap::getForeignURL() should be checked in case it is false, to avoid a CSRF DDoS via a malicious wikiid or notifywiki parameter.

Patch-set: 7
Label: Code-Review=-1