Update patch set 12
Patch Set 12:

Is it possible to use a longer token expiry, say 60s to match the squid timeout? I think 15s could be too short for some kinds of connections.

Is it possible to avoid displaying errors in the context of the central wiki? In particular, the memcached key could quite easily go missing due to server issues, resulting in centralauth-error-badtoken being delivered. The user is not expecting to be sent to another wiki when they click "log in", so being stuck there with all the skin links changed could be annoying.

The "return to the previous page" link goes some way towards addressing the navigational issue with errors in the central context, except that if the previous page was a response to a POST, clicking the "return" link in Firefox just goes to a "document expired" error message, which is disconcerting. It seems to me that just providing the error message name to the original wiki in a URL parameter would be a more consistent strategy.

Patch-set: 12
Reviewer: Tim Starling <17@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0