Update patch set 8
Patch Set 8:

I don't think the UI on Special:CentralLogin/complete is as good as it could be. I think there should be a spinner or progress bar to indicate that background login requests are taking place, with an icon or message indicating completion once the logins are done. I don't think replacing parts of the page text while the user is meant to be reading them is ideal.

Patch-set: 8
Label: Verified=0