Make CentralAuth login more silent
Change the redirect-through-central-domain to a 302 instead of a
JavaScript post (I'm told Safari 5 users aren't a concern) and make
the iframes load silently on the pageview after the login rather
than on a login success screen.

This adds the feature with a feature flag, which can be removed in a
future revision.