Update patch set 10
Patch Set 10: Code-Review+1

Looks like the recent patches addressed Aaron and Tim's comments.

I've tested this locally with all combinations of $wgCentralAuthLoginWiki, $wgCentralAuthSilentLogin and $wgCentralAuthUseOldAutoLogin. All are logging in the user cross-wiki for SUL. The only odd combination is when both wgCentralAuthSilentLogin and wgCentralAuthUseOldAutoLogin are false, then the local wiki's icon doesn't show up on the login success page (both with and without javascript). But the user still gets a centralauth_Session cookie from the local wiki, so the SUL part is still working, just oddity in the interface. We shouldn't ever run that combo of flags in production, but thought I would mention it.

Patch-set: 10