Redirect to correct protocol in SUL2
Attempt to handle situations where:
* The wiki has wgSecureLogin enabled, but the user opts-out of SSL--
this works if the check to the central wiki is always redirected to ssl
(how the WMF does it). Otherwise the central cookie is set encrypted
only, and the AutoLogin redirects happen over http, so the central wiki
thinks the user isn't logged in.
* The user logs in over http, and the central wiki forces the use of
SSL (e.g., how the WMF does it). In this case, the user is returned to
HTTP on the attached wiki, and they should be logged in for both secure
and unencrypted connections.

bug: 50334

Change-Id: I9e857ae265179922caa287de7ec65b14a5ebb83c