Allow extensions to change the SUL redirect
This may not be the best way to do it, but just a proposal to get the
conversation started.

This keeps things super simple, but it does mean an extension that
wants to redirect on things other than signup would need to have hooks
for both global users (and do the redirect with this hook) and non-
global users, and inject their content on the login success page.

Change-Id: I605cf01abb1838cd6dbd74314f282b256e8e17b8