Allow local wiki to act as central wiki
During testing / rollout, if a wiki gets a request to do the central
login, but doesn't have $wgCentralAuthLoginWiki set, it will redirect
back to the local wiki. The local wiki is supposed to take over, and
pretend to be the central wiki for handling the central wiki's part of
the login handshake.

This fixes a regression that seems to have shown up in wmf9?

Change-Id: Ife44d2856f6116f410b8f0cfe33a39f865a10d35