Better handle when $wgCentralAuthLoginWiki is not set on all wikis
During testing / rollout, if a wiki gets a request to do the central
login, but doesn't have $wgCentralAuthLoginWiki set, it will redirect
back to the local wiki, expecting it to act as a de-facto central wiki. But if the local wiki does have $wgCentralAuthLoginWiki set, then it will complain that it is not the central wiki.

Handle this situation by having the local wiki redirect to the central wiki.

Change-Id: Ife44d2856f6116f410b8f0cfe33a39f865a10d35