Allow caching of Special:CentralAutoLogin when safe
Special:CentralAutoLogin/start and
Special:CentralAutoLogin/checkLoggedIn are safe to cache, because
they only differ based on the session cookies (which the cache already
varies on) and don't have any side effects.

The other methods in the login flow do need to be uncached, though, as
they do have side effects.

Change-Id: Ie17b277857529f00a659cdbc933ff6c31fb616ed