Fix central session usage in Special:CentralAutoLogin
The central session expires when the user's browser is closed.
Special:CentralAutoLogin should check for this, and take appropriate
measures if the keys it had stored in the session are no longer present.

This fixes various PHP warnings and makes it better about honoring the
"remember me" checkbox from the original login.

Change-Id: I1f551fed312af7e70c5899d92e7f5726710f6f6d