Update patch set 1
Patch Set 1:

No, didn't bother filing a bug. The problem is that centralauth_Session (on all wikis, but particularly loginwiki) expires at the end of the browser session, and that's the only place we check for the saved "remember me" setting.

You can test it like this:
1. Login, with "remember me" checked.
2. Clear cookies for enwiktionary.
3. Visit enwiktionary. The JS check should work. If you check the cookies, centralauth_Token is set.
4. Close and reopen the browser.
5. Visit enwiktionary. You should still be logged in, thanks to centralauth_Token.
6. Clear cookies for enwiktionary.
7. Visit enwiktionary. The JS check should work. But if you check the cookies, centralauth_Token isn't set. Also, you'll see PHP undefined index warnings in your logs.
8. Close and reopen the browser.
9. Visit enwiktionary. You're not logged in (because no centralauth_Token), but the JS check should work same as in step 7.

The bug is step 7, the JS check didn't set centralauth_Token because it lost the "remember me" setting. It's visible in step 9, where you'd expect to already be logged in (like in step 5) but you're not.

Patch-set: 1
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0