Update patch set 2
Patch Set 2:

> it will only do that if MediaWiki tells it to using the X-Vary-Options and Vary headers. Whether this code indeed makes sure of that for these URLs, I haven't checked.

The call to OutputPage's sendCacheControl does it. That function forces Vary to contain Cookie, and the default it uses for X-Vary-Options will vary on the session cookie and other login cookies (it's possible something on the 'GetCacheVaryCookies' hook could delete those from X-Vary-Options, but if it does that then it's breaking caching all over the place).


Also, I just realized we can cache /createSession if the passed gu_id is 0, so PS2 makes that happen.

Patch-set: 2
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0