Delete forceHTTPS cookies on logout
To do this, we need to resurrect the icons on the logout page so that
the cookies can be removed on all the other domains.

We also need to take care of the forceHTTPS cookie set by core.
Otherwise say the user logs in on dewiki, they wind up with two
forceHTTPS cookies (one for de.wikipedia.org set by core and one for
.wikipedia.org set by CentralAuth). If they later log out on some other
site, only the latter will be deleted.

Bug: 53536
Change-Id: I57c7f3dd3b040ad4d4fc9ec15ec13eddbfb781e4