Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

We need some way to make sure only the user can initiate this for their account (an attacker wouldn't get anything by csrfing a user and logging them out, but it would be annoying). Probably generate a token, like in the old-style loggout, add their username to memcache, and do the loggout on the memcached name. Also, set a flag on the memcache object to make sure this was a logout action.

Patch-set: 1
Label: Code-Review=-1