Update patch set 1
Patch Set 1:

(2 comments)

I did the memc thing, and then I realized it doesn't make a bit of difference since there's still no way to validate the passed username. Normally we'd not pass the name at all and just get it from $user->getName(), but by the time this is called the user should already be logged out so we can't even validate it that way.

Actually, why don't we do just that to prevent this attack: refuse to delete the cookies if the user is still logged in. And for good measure, we can get the username from the still-existing cookie that we're about to delete, instead of passing it at all.

Patch-set: 1
Reviewer: Anomie <221@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0