Update patch set 2
Patch Set 2:

Ah, sorry, I wasn't thinking it all the way through either.

I think it would be good to validate that the session is invalid (like you're doing), but instead of checking that the User cookie is for a valid user and the user is attached, compare that with something passed in the request (or pass a token in the request that points to a memcache object that has the username) and make sure the usernames match.

Patch-set: 2
Label: Verified=0