Re-do edge login when JS login check succeeds
People are reporting that they're still sometimes logged out on some
wikis, even when the JS login check succeeds. It has been requested that
we re-do the edge login pixel images in that situation to try and fix
this issue.

Also, to reduce the incidence of skipped auto-login checks, let's add a
small RL module to all logged-in pageviews to clear CentralAuthAnon.

Also, to avoid race conditions (e.g. user visits enwiki and frwiki from
bookmarks and both fire the JS login check, but the second "fails"
because the first already succeeded) let's not fail the auto-login just
because we're already logged in; this might reset edit tokens in some
rare cases, but that only causes the "session data lost, try again"
screen people already see if they leave the edit screen sitting for too
long.

Bug: 53953
Bug: 54147
Change-Id: I418ac068c0856c828235c0b39f1c1f6cfcaebb2d