Re-do edge login when JS login check succeeds
People are reporting that they're still sometimes logged out on some
wikis, even when the JS login check succeeds. It has been requested that
we re-do the edge login pixel images in that situation to try and fix
this issue.

Note this still isn't foolproof, since e.g. if somehow enwiki is logged
in while other-language Wikipedias aren't, the edge login will report
"already logged in".

Bug: 53953
Change-Id: I418ac068c0856c828235c0b39f1c1f6cfcaebb2d