Set Squid maxage so proxies will cache
Since OutputPage will send Cache-Control: private if s-maxage is 0
(even with mEnableClientCache set, see OutputPage::sendCacheControl())
set a squid maxage for /start and /checkLoggedIn phases.

The proxy *must not* use the cached value if the client is sending
cookies with the request, so Vary headers are also set for the call
to loginwiki (/checkLoggedIn).

Bug: 54195
Change-Id: I45e2167826d91e6132a35c3fab99843f481e1c5a