Update patch set 1
Patch Set 1:

This make the caching headers more accurate, but I think we actually can't cache this redirect at all since we also essentially vary on the remote ip when we do wfCanIPUseHTTPS( $request->getIP() ). So someone using http from china will not be redirected, while someone using http from the US will be redirected.

These particular redirect should  be very rare, since that would mean the user had either submitted their initial login to an http url, but wfCanIPUseHTTPS is true and we want them to use https for the central login (I'm not sure how that case would be triggered on the cluster). The second way is the autologin, which would be hit for users who have opted out of https for their logged in session, and then hit a wiki where they aren't logged in, so we do the auto login originating from a logged in http connection. That case should be rare.

Patch-set: 1
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0