Update patch set 1
Patch Set 1:

The reason I initially didn't allow self-renames was because the user will be forced logged out, which could cause issues if that was the user executing the action. This is still a WIP so I'll look at it later down the road to see if it in fact is an actual issue.

Patch-set: 1
Label: Verified=0