Update patch set 2
Patch Set 2: Code-Review+2

This is working under several different scenarios of wgSecureLogin that have in my dev environment. The resource loader bit should be fully cacheable now, and there shouldn't be any issues if this is rolled out to the wikis with the normal deployment schedule, since it's compatible with the current version of CentralAuth.

This should eliminate the calls to /start. The calls to /checkLoggedIn can be served from the cache when none of the logged in cookies are present.

Merging this now, and I'll let QA know to watch out for issues in beta.

Patch-set: 2
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2