Changed Special:GlobalUsers to inherit from AlphabeticPager
Previously, Special:GlobalUsers inherited from UsersPager, which is used to
display Special:ListUsers. The problem with that was that not all of the
controls generated by UsersPager are applicable to GlobalUsers, so some of
them were useless - specifically the "Show only users with edits" and
"Sort by creation date" checkboxes. This addresses bug 57544, and a fix for
this bug is especially useful to allow the implementation of a fix for
bug 33545, which introduces new controls to Special:ListUsers.

Bug: 57544
Change-Id: I53b7b7a273152ed80b71f58c7b5d1c28c1e5c70c