Update patch set 2
Patch Set 2:

Duplicate if there is a fair chance the strings will be customised or have a different context. I the context is exactly the same, you can re-use.

I'm missing a rationale for using a higher level base class to extend. Only thing I see it that it adds 60 lines of code... Can this be added to the commit summary?

Patch-set: 2