Don't localize Special:CentralAutoLogin
Localization of Special:CentralAutoLogin causes problems for log
analysis and for dealing with the large number of hits that can come
into this page (e.g. bug 54195 comment 20). At the same time, it gives
miniscule benefit to users, as it is never exposed in the public user
interface.

So let's just remove the localization of it. To really make it work, we
also need to prevent the redirect from "Special:CentralAutoLogin" to the
localized version of the "Special" namespace.

Bug: 54195
Change-Id: I820370de3afa0bdfce360cba1aef651ac084fd06