Correct failure message when account is locked
When the account is globally locked, CentralAuth will currently fail the
password check leading to a message "Incorrect password entered." This
is misleading and causes unnecessary bug reports such as bug 53755 and
bug 57791.

CentralAuth should instead produce a message informing the user that
their account is globally locked, along the lines of 'login-userblocked'
used when $wgBlockDisablesLogin is set and the user is locally blocked.

Bug: 57866
Change-Id: I5674ed88a002cb9a0e0b09e040450f75906e058a