Update patch set 1
Patch Set 1:

It could be my dev env, but it looks like all users get the edit count of the first user on the list.

When logged the output of $globalUser->queryAttached() when building the row, and I'm getting the same set of local users for each global user. Which is odd...

Ah, fun queryAttached() keeps a static of the results, assuming you'll never have multiple CentralAuthUsers I guess.

Patch-set: 1
Label: Code-Review=0