Update patch set 47
Patch Set 47:

@Nemo, it's not the end of the world, but if they do any edits, then the logs/revisions might have the old name (depending on the timing). So those would need an extra manual cleanup.

If we first check renameInProgressOn(wfWikiID()) in onAbortLogin (which is relative rare, compared to loading from session), then we can always abort (I think) if either the to or from name is being renamed, since that would only return something if there was a rename for a global account on this wiki that hadn't run. After that, we still have to do the renamInProgress() check and abort if it's attached (to catch user logging in with the new name).

Patch-set: 47
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0