Update patch set 49
Patch Set 49:

PS48: Chris updated the AbortLogin hook based on his earlier comment. It now adds a query to master on login, but that is rare (compared to on every page view), so it should be acceptable.
PS49-50: I moved the query in CentralAuthUser::loadState out of the join, so it now works for the oldname. It requires a separate query, but it will be cached. Also "fixed" CentralAuthUserArray to also check the renameuser_status table.

Patch-set: 49
Label: Verified=0