Update patch set 11
Patch Set 11: Code-Review-1

Generally, this looks really good—your approach to writing the tests is spot on. I've just tidied up some dependencies and put in an assertion for the presence of the required env variables.

One small thing about your assertions: Use `expect(subject).to matcher` syntax rather than `.should`. The latter is deprecated in rspec 2.99 and going away in 3.0.[1]

The login.feature fails for me in the mw-vagrant env. This may just be a configuration issue, but I didn't have time to dig in further. I've pushed a commit for the necessary mw-vagrant setup so you can repro it.[2]

1: https://relishapp.com/rspec/rspec-expectations/docs
2: If2aefdabc513de332901d09bf7b2c2699a8932ba

Patch-set: 11
Label: Code-Review=-1