[WIP] Interstitial notification page for renamed users
Redirect users who have been authenticated to a renamed central account
using their previous local username to an interstital page telling them
to login using the new name in the future and giving other helpful
advice about the SUL rename process.

Content for the interstitial page is not final and should be considered
to be placeholder text for now.

Bug: 67995
Change-Id: I7bf1dd18d578b52e29e2f29933fcbd9b3261ae99