Update patch set 9
Patch Set 9: Code-Review-1

I tested this out yesterday, and it looks like we're still not handling the case where the non-renamed account doesn't exist. So if foo is renamed to foo~enwiki, but then the winner (foo) is renamed to something else, this still fails during the local part because in SpecialUserlogin::attemptAutoCreate, $wgAuth->userExists( $user->getName() ) returns false.

I'm not sure the best way to get around that..

Otherwise things look good.

Patch-set: 9
Label: Code-Review=-1